REGRESSION(r216977): [GTK] Ephemeral sessions broken after r216977 https://bugs.webkit.org/show_bug.cgi?id=172345
Reviewed by Michael Catanzaro.

Since r216977, WebProcessPool sends the page session ID to the web process, but the WebsiteDataStoreParameters
session ID to the network process. WebsiteDataStoreParameters is only implemented for Cocoa, so we are sending 0
session ID to the network process, while using the right session ID everywhere else.

* UIProcess/WebProcessPool.cpp:
(WebKit::WebProcessPool::pageAddedToProcess): Add an ASSERT to ensure we are sending the same session ID to Web
and Network processes.
* UIProcess/WebsiteData/WebsiteDataStore.cpp:
(WebKit::WebsiteDataStore::parameters): Add minimal implementation for non-Cocoa ports to at least set the
session ID.

Identifier: 189263@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@217128 268f45cc-cd09-0410-ab3c-d52691b4dbfc