REGRESSION(r217051): Automation sessions fail to complete bootstrap https://bugs.webkit.org/show_bug.cgi?id=172513 <rdar://problem/32338354>
Reviewed by Joseph Pecoraro.

The changes to be more strict about typechecking messages were too strict.

* inspector/remote/cocoa/RemoteInspectorCocoa.mm:
(Inspector::RemoteInspector::receivedSetupMessage):
WIRAutomatically is an optional key in the setup message. In the relay, this key gets copied
into an NSDictionary as NSNull if the key isn't present in a forwarded command.
We need to revert NSNull values to nil, since it's valid to call [nil boolValue] but not
[[NSNull null] boolValue]. We also need to allow for nil in the typecheck for this key.

Identifier: 189430@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@217309 268f45cc-cd09-0410-ab3c-d52691b4dbfc