REGRESSION (r206481): Don't assume frameCount() is larger than or equal to the size of the image frame cache https://bugs.webkit.org/show_bug.cgi?id=172552
Patch by Said Abou-Hallawa <sabouhallawa@apple.com> on 2017-05-24
Reviewed by Geoffrey Garen.

If an image with many cached ImageFrames is cached but its encoded data is
freed, then the same image is reused, the encoded data will be fetched.
But the initial frameCount() might be less than the current size of the
ImageFrames cache.

This was causing an assertion in the debug build and crash in the release
build. Vector::grow() does not handle the case where the new size is smaller
than the current size safely.

* platform/graphics/ImageFrameCache.cpp:
(WebCore::ImageFrameCache::growFrames):

Identifier: 189467@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@217392 268f45cc-cd09-0410-ab3c-d52691b4dbfc