REGRESSION(216914): testCFStrings encounters an invalid ExecState callee pointer. https://bugs.webkit.org/show_bug.cgi?id=172651
Reviewed by Saam Barati.

This is because the assertion utility functions used in testCFStrings() expects
to get the JSGlobalContextRef from the global context variable.  However,
testCFStrings() creates its own JSGlobalContextRef but does not set the global
context variable to it.

The fix is to make testCFStrings() initialize the global context variable properly.

* API/tests/testapi.c:
(testCFStrings):



Identifier: 189553@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@217498 268f45cc-cd09-0410-ab3c-d52691b4dbfc