[Mac] Use compositing for frames when ScrollingTreeIncludesFrames=true https://bugs.webkit.org/show_bug.cgi?id=172851
Patch by Frederic Wang <fwang@igalia.com> on 2017-06-02
Reviewed by Simon Fraser.

Source/WebCore:

When fast frame scrolling is supported on mac (bug 171667), scrollable frames will require
special composited layer. This commit enables that on mac but only when
ScrollingTreeIncludesFrames=true, so that the default behavior is unchanged.

Test: compositing/iframes/compositing-for-scrollable-iframe.html

* rendering/RenderLayerCompositor.cpp:
(WebCore::RenderLayerCompositor::requiresCompositingForScrollableFrame): Do not enable frame
compositing when ScrollingTreeIncludesFrames=false.
Rename ScrollableInnerFrameTrigger to ScrollableNonMainFrameTrigger.
* page/ChromeClient.h: Rename ScrollableInnerFrameTrigger.

Source/WebKit2:

When fast frame scrolling is supported on mac (bug 171667), scrollable frames will require
special composited layer. This commit enables that on mac but only when
ScrollingTreeIncludesFrames=true, so that the default behavior is unchanged.

* WebProcess/WebCoreSupport/WebChromeClient.h: Make scrollable inner frame trigger
compositing on Mac by default. This is ignored on mac when ScrollingTreeIncludesFrames=false.
Rename ScrollableInnerFrameTrigger to ScrollableNonMainFrameTrigger.

LayoutTests:

This commit adds a test to check whether scrollable iframes require a special composited
layer when ScrollingTreeIncludesFrames=true. This is currently only done on Mac WK2.

* compositing/iframes/compositing-for-scrollable-iframe.html: Added. This is a basic test
printing the layer tree of a page with an iframe.
* platform/ios-simulator/compositing/iframes/compositing-for-scrollable-iframe-expected.txt: Added.
* platform/mac/compositing/iframes/compositing-for-scrollable-iframe-expected.txt: Added.
* platform/mac-wk1/compositing/iframes/compositing-for-scrollable-iframe-expected.txt: Added.

Identifier: 189754@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@217726 268f45cc-cd09-0410-ab3c-d52691b4dbfc