[Mac] Include frames in the scrolling tree when ScrollingTreeIncludesFrames=true https://bugs.webkit.org/show_bug.cgi?id=172825
Patch by Frederic Wang <fwang@igalia.com> on 2017-06-02
Reviewed by Simon Fraser.

Source/WebCore:

This commit generalizes RenderLayerCompositor/ScrollingCoordinator so that it handles the
case of non-main frames. When ScrollingTreeIncludesFrames=true, all the frames will appear
in the scrolling tree on Mac WK2.

Test: fast/scrolling/scrolling-tree-includes-frame.html

* page/scrolling/ScrollingCoordinator.cpp:
(WebCore::ScrollingCoordinator::coordinatesScrollingForFrameView): Do not to restrict to the
main frame.
* rendering/RenderLayerCompositor.cpp:
(WebCore::RenderLayerCompositor::updateBacking): Do not to restrict to the main frame.
(WebCore::RenderLayerCompositor::useCoordinatedScrollingForLayer): Make this function a class
member so it can call hasCoordinatedScrolling and no longer needs a RenderView parameter.
(WebCore::RenderLayerCompositor::updateScrollCoordinatedStatus): Do not to restrict to the
main frame.
* rendering/RenderLayerCompositor.h: Introduce useCoordinatedScrollingForLayer.

LayoutTests:

Add a new test to check whether a non-main frame appears in the scrolling tree when
ScrollingTreeIncludesFrames=true. For now this is only happens for Mac WK2. Also update
the iOS WK2 reference for scrolling-iframe-expected.html

* fast/scrolling/scrolling-tree-includes-frame-expected.txt: Added.
* fast/scrolling/scrolling-tree-includes-frame.html: Added.
* platform/ios-simulator/fast/scrolling/scrolling-tree-includes-frame-expected.txt: Added.
* platform/ios-wk2/compositing/iframes/scrolling-iframe-expected.txt: Update reference.
* platform/mac-wk1/fast/scrolling/scrolling-tree-includes-frame-expected.txt: Added.

Identifier: 189758@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@217730 268f45cc-cd09-0410-ab3c-d52691b4dbfc